Rename DisplayInterface file to DisplayAbstract
