+++ Platformio/platformio.ini
@@ -48,7 +48,7 @@ lib_archive = false
 build_src_filter =
   +<../OmoteUI/*>
   +<../HAL/HardwareAbstract.cpp>
-  +<../HAL/DisplayInterface.cpp>
+  +<../HAL/HardwareInterfaces/DisplayAbstract.cpp>
